Fix some compiler warnings

+++ Libraries/lwip_v1.3.2/port/STM32F4x7/arch/cc.h
@@ -38,8 +38,8 @@ typedef unsigned   char    u8_t;
 typedef signed     char    s8_t;
 typedef unsigned   short   u16_t;
 typedef signed     short   s16_t;
-typedef unsigned   long    u32_t;
-typedef signed     long    s32_t;
+typedef unsigned   int     u32_t;
+typedef signed     int     s32_t;
 typedef u32_t mem_ptr_t;
 typedef int sys_prot_t;
